+++ tests/test_objects_variants.py
@@ -2171,3 +2171,81 @@ def test_get_operations_help():
     print(actual_operations_help[0])
 
     assert expected_operations_help[0] == actual_operations_help[0]
+
+
+def test_rename_fields():
+    """
+    The function `test_rename_fields` renames specified fields in a VCF file and checks if the output
+    VCF is in the correct format using pyVCF.
+    """
+
+    with TemporaryDirectory(dir=tests_folder) as tmp_dir:
+
+        # Init files
+        input_vcf = tests_data_folder + "/example.annotation_names.vcf"
+        output_vcf = f"{tmp_dir}/output.vcf"
+
+        # Create object
+        variants = Variants(input=input_vcf, load=True)
+
+        # Fieldst to rename
+        fields_to_rename = {
+            "CLNSIG": "CLNSIG_renamed",
+            "PREFIXCLNSIG": "PREFIXCLNSIG_renamed",
+            "DP": "depth",
+            "field_not_in_header": "field_not_in_header_renamed",
+            "": "",
+            "SIFT": None,
+        }
+
+        # Rename fields
+        fields_renamed = variants.rename_fields(fields_to_rename=fields_to_rename)
+        assert fields_renamed == {'CLNSIG': 'CLNSIG_renamed', 'PREFIXCLNSIG': 'PREFIXCLNSIG_renamed', 'DP': 'depth', 'SIFT': None}
+
+        # Check if VCF is in correct format with pyVCF
+        remove_if_exists([output_vcf])
+        variants.export_output(output_file=output_vcf)
+        try:
+            vcf_obj = vcf.Reader(filename=output_vcf)
+        except:
+            assert False
+
+
+def test_rename_fields_to_param_and_export():
+    """
+    The function `test_rename_fields_to_param_and_export` renames specified fields in a VCF file and
+    checks if the output VCF is in the correct format using pyVCF.
+    """
+
+    with TemporaryDirectory(dir=tests_folder) as tmp_dir:
+
+        # Init files
+        input_vcf = tests_data_folder + "/example.annotation_names.vcf"
+        output_vcf = f"{tmp_dir}/output.test.vcf"
+
+        # Param
+        param = {
+            "export": {
+                "fields_to_rename": {
+                    "CLNSIG": "CLNSIG_renamed",
+                    "PREFIXCLNSIG": "PREFIXCLNSIG_renamed",
+                    "DP": "depth",
+                    "field_not_in_header": "field_not_in_header_renamed",
+                    "": "",
+                    "SIFT": None,
+                }
+            }
+        }
+
+        # Create object
+        variants = Variants(input=input_vcf, param=param, load=True)
+
+        # Check if VCF is in correct format with pyVCF
+        remove_if_exists([output_vcf])
+        variants.export_output(output_file=output_vcf)
+        try:
+            vcf_obj = vcf.Reader(filename=output_vcf)
+            assert list(set(vcf_obj.infos.keys())).sort() == ['CLNSIGSUFFIX', 'AA', 'NS', 'PREFIXCLNSIG_renamed', 'CLNSIG_renamed', 'depth'].sort()
+        except:
+            assert False
+